bus: fix capabilities on big-endian
The kernel provides capabilities as a u32 array, sd-bus uses an u8 array.
This works fine on little-endian as both are encoded the same way.
However, this fails on big-endian if we do not perform sufficient
byte-swapping on each u32 entry.

This patch makes sd-bus use u32, too. We avoid changing any kernel
provided data so we can keep pointing into kdbus pool buffers which
contain u32 arrays.

+++ src/libsystemd/sd-bus/bus-creds.h
@@ -60,7 +60,7 @@ struct sd_bus_creds {
         char *user_unit;
         char *slice;
 
-        uint8_t *capability;
+        uint32_t *capability;
 
         uint32_t audit_session_id;
         uid_t audit_login_uid;
